+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v1.6.5-rc2 - ????/??/??
 
 - [FEATURE] Add support for new reCAPTCHA version in `AntiBot` plugin
+- [ALL-IN-ONE] Add support for disabling specific CrowdSec parsers
+- [ALL-IN-ONE] Update CrowdSec version to 1.7.0
 - [DOCS] Add multi-language support to the documentation, including French
 - [DEPS] Update lua-resty-session version to v4.1.4
 - [DEPS] Update lua-resty-openssl version to v1.6.2